:tractor: code refactoring and update :books:
